Revert to plaintext
I'll probably go to hell for this...

but I need to show the password to the user on the site, so either
reversable encryption, or plaintext. Plaintext because the password will
be a x-character long random generated password which does not contain
sensitive user information, and is hence useless to a hacker in that
way.

:hear_no_evil: